requirements.txt
Flask
PyGithub
transformers
torch
pylint
eslint
